+++ .gitignore
@@ -165,3 +165,6 @@ dist
 .ionide
 
 # End of https://www.toptal.com/developers/gitignore/api/visualstudiocode,node
+
+### Agents ###
+.agents
