+++ editor/index.html
@@ -39,23 +39,23 @@
     <script async src="https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js?client=ca-pub-6795302641432113"
         crossorigin="anonymous"></script>
     <meta charset="UTF-8">
-    <meta property="og:type" content="website" />
-    <meta property="og:site_name" content="Easy Discord Bot Builder" />
-    <meta property="og:title" content="Easy Discord Bot Builder" />
-    <meta property="og:description" content="DiscordのBotをノーコードで。" />
-    <meta property="og:url" content="https://himais0giiiin.com//" />
-    <meta property="og:image" content="https://himais0giiiin.com/static/ogp.jpg" />
-    <meta property="og:locale" content="ja_JP" />
+    <meta property="og:type" content="website">
+    <meta property="og:site_name" content="Easy Discord Bot Builder (EDBB)">
+    <meta property="og:title" content="Easy Discord Bot Builder (EDBB)">
+    <meta property="og:description" content="ブラウザ上でブロックを組み合わせて、誰でも簡単に Discord Bot を作成できるビジュアルプログラミングツール。全て無料。">
+    <meta property="og:url" content="https://himais0giiiin.com/editor/">
+    <meta property="og:image" content="https://himais0giiiin.com/static/ogp.jpg">
+    <meta property="og:locale" content="ja_JP">
 
-    <meta name="twitter:card" content="summary_large_image" />
-    <meta name="twitter:title" content="Easy Discord Bot Builder" />
-    <meta name="twitter:description" content="DiscordのBotをノーコードで。" />
-    <meta name="twitter:image" content="https://himais0giiiin.com/static/ogp.jpg" />
+    <meta name="twitter:card" content="summary_large_image">
+    <meta name="twitter:title" content="Easy Discord Bot Builder (EDBB)">
+    <meta name="twitter:description" content="DiscordのBotをノーコードで。">
+    <meta name="twitter:image" content="https://himais0giiiin.com/static/ogp.jpg">
 
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
 
 
-    <title>Easy Discord Bot Builder</title> <!-- nは更新したときに見やすくするものです -->
+    <title>Easy Discord Bot Builder</title>
 
     <link rel="icon" type="image/svg+xml" href="../static/icon.svg">
 
@@ -103,8 +103,8 @@
     <script src="https://unpkg.com/blockly@11.2.2/python_compressed.js"></script>
     <script src="https://unpkg.com/blockly@11.2.2/msg/ja.js"></script>
     <script src="https://unpkg.com/@blockly/field-multilineinput@4.0.4/dist/index.js"></script>
-    <script src="./mobile.js"></script>
-    <script src="./plugin-mobile.js"></script>
+    <script type="module" src="./mobile.js"></script>
+    <script type="module" src="./plugin-mobile.js"></script>
     <link rel="stylesheet" href="./mobile.css">
     <!-- Icons -->
     <script src="https://unpkg.com/lucide@latest"></script>
@@ -132,12 +132,6 @@
 
 
     <link rel="stylesheet" href="style.css">
-    <meta property="og:title" content="Easy Discord Bot Builder (EDBB)">
-    <meta property="og:site_name" content="Easy Discord Bot Builder(EDBB)">
-    <meta property="og:type" content="website">
-    <meta property="og:url" content="https://himais0giiiin.com/beta/">
-    <meta property="og:description" content="ブラウザ上でブロックを組み合わせて、誰でも簡単に Discord Bot を作成できるビジュアルプログラミングツール。全て無料。">
-    <meta property="og:image" content="https://himais0giiiin.com/static/img.svg">
 </head>
 
 <body
@@ -170,14 +164,13 @@
                 </div>
                 <div
                     class="relative bg-white dark:bg-slate-900 p-2.5 rounded-xl shadow-sm border border-slate-100 dark:border-slate-800 flex items-center justify-center">
-                    <img src="../static/icon.svg" alt="EDBB" class="w-6 h-6" style="margin-top: 1px;">
+                    <img src="../static/icon.svg" alt="EDBB" class="w-6 h-6 mt-[1px]">
                 </div>
             </div>
             <div class="flex flex-col">
                 <h1
                     class="text-xl font-bold text-slate-900 dark:text-white leading-tight tracking-tight flex items-center gap-3">
                     Easy Discord Bot Builder
-                    <!-- <span class="text-[10px] px-2 py-0.5 rounded-full bg-gradient-to-r from-indigo-500/10 to-purple-500/10 border border-indigo-500/20 text-indigo-600 dark:text-indigo-400 font-mono font-bold">BETA</span> -->
                 </h1>
                 <p class="text-[11px] font-medium text-slate-500 dark:text-slate-400 mt-0.5 flex items-center gap-1">
                     Create by himais0giiiin!
@@ -224,24 +217,24 @@
                 <button id="importBtn"
                     class="p-2 text-slate-500 hover:text-indigo-600 dark:text-slate-400 dark:hover:text-indigo-400 transition-colors rounded-md hover:bg-slate-50 dark:hover:bg-slate-800"
                     title="Load JSON">
-                    <i data-lucide="upload" class="w-4.5 h-4.5"></i>
+                    <i data-lucide="upload" class="w-5 h-5"></i>
                 </button>
                 <button id="exportBtn"
                     class="p-2 text-slate-500 hover:text-indigo-600 dark:text-slate-400 dark:hover:text-indigo-400 transition-colors rounded-md hover:bg-slate-50 dark:hover:bg-slate-800"
                     title="Save JSON">
-                    <i data-lucide="download" class="w-4.5 h-4.5"></i>
+                    <i data-lucide="download" class="w-5 h-5"></i>
                 </button>
                 <button id="jsonGuiBtn"
                     class="p-2 text-slate-500 hover:text-indigo-600 dark:text-slate-400 dark:hover:text-indigo-400 transition-colors rounded-md hover:bg-slate-50 dark:hover:bg-slate-800"
                     title="JSON GUIを開く">
-                    <i data-lucide="table-2" class="w-4.5 h-4.5"></i>
+                    <i data-lucide="table-2" class="w-5 h-5"></i>
                 </button>
 
                 <div data-theme-divider class="w-px h-4 bg-slate-200 dark:bg-slate-700 mx-1"></div>
                 <button id="themeToggle"
                     class="p-2 text-slate-500 hover:text-amber-500 dark:text-slate-400 dark:hover:text-amber-400 transition-colors rounded-md hover:bg-slate-50 dark:hover:bg-slate-800">
-                    <i data-lucide="moon" class="hidden w-4.5 h-4.5 dark:block"></i>
-                    <i data-lucide="sun" class="block w-4.5 h-4.5 dark:hidden"></i>
+                    <i data-lucide="moon" class="hidden w-5 h-5 dark:block"></i>
+                    <i data-lucide="sun" class="block w-5 h-5 dark:hidden"></i>
                 </button>
             </div>
 
@@ -1249,9 +1242,7 @@
                 <div
                     class="p-3 rounded-xl border border-blue-200 dark:border-blue-900/50 bg-blue-50/70 dark:bg-blue-900/10 flex items-center gap-2">
                     <i data-lucide="monitor" class="w-4 h-4 text-blue-600 dark:text-blue-400 shrink-0"></i>
-                    <p class="text-xs font-medium text-blue-900 dark:text-blue-100">
-                        Windows / Linux で利用できます
-                    </p>
+                    <p class="text-xs font-medium text-blue-900 dark:text-blue-100">Windows / Linux で利用できます</p>
                 </div>
 
                 <div
@@ -1336,7 +1327,6 @@
         }
     </script>
     <script type="module" src="script.js?v=20260228a"></script>
-    <script type="module" src="./src/main.ts"></script>
     <style>
         #docsbotai-root {
             position: fixed !important;
