+++ index.html
@@ -2,18 +2,18 @@
 <html lang="ja">
 
 <head>
-    <meta property="og:type" content="website" />
-    <meta property="og:site_name" content="Easy Discord Bot Builder" />
-    <meta property="og:title" content="Easy Discord Bot Builder" />
-    <meta property="og:description" content="DiscordのBotをノーコードで。" />
-    <meta property="og:url" content="https://himais0giiiin.com//" />
-    <meta property="og:image" content="https://himais0giiiin.com/static/ogp.png" />
-    <meta property="og:locale" content="ja_JP" />
+    <meta property="og:type" content="website">
+    <meta property="og:site_name" content="Easy Discord Bot Builder">
+    <meta property="og:title" content="Easy Discord Bot Builder">
+    <meta property="og:description" content="DiscordのBotをノーコードで。">
+    <meta property="og:url" content="https://himais0giiiin.com/">
+    <meta property="og:image" content="https://himais0giiiin.com/static/ogp.png">
+    <meta property="og:locale" content="ja_JP">
 
-    <meta name="twitter:card" content="summary_large_image" />
-    <meta name="twitter:title" content="Easy Discord Bot Builder" />
-    <meta name="twitter:description" content="DiscordのBotをノーコードで。" />
-    <meta name="twitter:image" content="https://himais0giiiin.com/static/ogp.png" />
+    <meta name="twitter:card" content="summary_large_image">
+    <meta name="twitter:title" content="Easy Discord Bot Builder">
+    <meta name="twitter:description" content="DiscordのBotをノーコードで。">
+    <meta name="twitter:image" content="https://himais0giiiin.com/static/ogp.png">
 
     <meta charset="UTF-8">
     <meta name="viewport"
